In a completely devastated world like this...there isn't anyone left who hasn't lost someone important.
