+++ fractional-NFT/project/contracts/fractional-NFT-contract/tests/functions/core/mod.rs
@@ -1,0 +1,3 @@
+mod deposit;
+mod set_admin;
+mod withdraw;
