+++ fractional-NFT/project/contracts/fractional-NFT-contract/tests/functions/info/mod.rs
@@ -1,0 +1,2 @@
+mod nft_info;
+mod supply;
