+++ fractional-NFT/project/contracts/fractional-NFT-contract/tests/functions/mod.rs
@@ -1,5 +1,2 @@
-mod deposit;
-mod nft_info;
-mod set_admin;
-mod supply;
-mod withdraw;
+mod core;
+mod info;
